optimize:move utils func to pkg
