+++ utils/hash/md5_test.go
@@ -12,17 +12,30 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package utils
+package hash
 
-import (
-	"os"
-	"path/filepath"
+import "testing"
 
-	"github.com/sealerio/sealer/common"
-)
-
-func ExecutableFilePath() string {
-	ex, _ := os.Executable()
-	exPath := filepath.Dir(ex)
-	return filepath.Join(exPath, common.ExecBinaryFileName)
+func TestMD5(t *testing.T) {
+	type args struct {
+		body []byte
+	}
+	tests := []struct {
+		name string
+		args args
+		want string
+	}{
+		{
+			"test md5",
+			args{body: []byte("test data")},
+			"eb733a00c0c9d336e65691a37ab54293",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := MD5(tt.args.body); got != tt.want {
+				t.Errorf("MD5() = %v, want %v", got, tt.want)
+			}
+		})
+	}
 }
