+++ utils/os/env.go
@@ -1,4 +1,4 @@
-// Copyright © 2021 Alibaba Group Holding Ltd.
+// Copyright © 2022 Alibaba Group Holding Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,34 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package utils
+package os
 
 import (
-	"crypto/md5" // #nosec
-	"encoding/hex"
 	"fmt"
-	"io"
 	"os"
-	"path/filepath"
 )
 
-func MD5(body []byte) string {
-	bytes := md5.Sum(body) // #nosec
-	return hex.EncodeToString(bytes[:])
-}
-
-//FileMD5 count file md5
-func FileMD5(path string) (string, error) {
-	file, err := os.Open(filepath.Clean(path))
-	if err != nil {
-		return "", err
-	}
-
-	m := md5.New() // #nosec
-	if _, err := io.Copy(m, file); err != nil {
-		return "", err
-	}
-
-	fileMd5 := fmt.Sprintf("%x", m.Sum(nil))
-	return fileMd5, nil
+func SetRootfsBinToSystemEnv(rootfs string) error {
+	bin := fmt.Sprintf(":%s/bin", rootfs)
+	return os.Setenv("PATH", os.Getenv("PATH")+bin)
 }
